LayoutTests/crypto/webkitSubtle/sha-1-expected.txt

CONSOLE MESSAGE: line 1: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test crypto.webkitSubtle.digest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


SHA1 of []
PASS bytesToHexString(new Uint8Array(digest)) is 'da39a3ee5e6b4b0d3255bfef95601890afd80709'
SHA1 of [0x0]
PASS bytesToHexString(new Uint8Array(digest)) is '5ba93c9db0cff93f52b521d7420e43f6eda2784f'
SHA1 of [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
PASS bytesToHexString(new Uint8Array(digest)) is '2c7e7c384f7829694282b1e3a6216def8082d055'
PASS crypto.webkitSubtle.generateKey('sha-1') threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS successfullyParsed is true

TEST COMPLETE
